SK:
- Made elastix much more memory-friendly.
- Bug in BaseComponentSE (forgot to check if a pointer is null or not).
 this bug didn't cause any trouble before, luckily. It would have caused
problems with this new memory-friendly elastix.
- More attention is paid to making sure that the Transform has stored a copy
of the final parameters (needed because it will be used in the next elastix
level). Especially was likely to give problems, because it doesn't make a real
copy of the parameters normally. it just stores the address of the data.
we were very lucky that we didn't have problems with this before....
it would have given problems with this new memory-friendly elastix.

NB: still have to check the deformationfieldtransform and the BSplineWithDiffusion
for the same error.