SK:
-BUG: major bug in FastAndLowMemory implementation. an Array2D
(m_PRatioArray) was not initialized with zeroes, while there was a
chance that some elements were not set later on. This caused very weird
crashes, almost not reproducible at all (adding a std::cout somewhere in
the code could already solve the problem). I fixed the initialization,
which fixed the crashes, but it could be that these crashes were not because of this bug, since
simply adding a std::cout also resolved the error :S. Anway, this was
for sure a bug, and had to be fixed.