MS:
-ENH: modified the WritePyramidImage such that it writes all pyramids, not just the first one.

The implementation currently has a bug, namely that the images are written as many times as there are pyramids, which should be just once. To solve this bug we need to make the BeforeEachResolutionBase(void) aware of for which pyramid it is called in ElastixTemplate->CallInEachComponent(). For that I think we need to change the signature of BeforeEachResolutionBase to e.g. BeforeEachResolutionBase( const unsigned int whichOneCalledIt ) or store a member variable or something. I will hold that off until we have discussed this.