MS:
-BUG: elastix would segfault when an invalid component was supplied in the parameter file, without giving any hints. For example:
  (MovingImagePyramid "MovingRecursiveImagePyramid" "FixedRecursiveImagePyramid")
The FixedRecursiveImagePyramid would simply be set in the moving pyramid, but at a later stage in ConfigureComponents from ElastixTemplate a dynamic cast would be performed to the base type. This dynamic cast (elxGetBaseMacro) was never checked for success, but simply used, which caused the segfault when an invalid type was specified.
Now I check for the success of this cast and throw a meaningful and informative error message.