MS:
ENH: set doxygen parameter SHORT_NAMES from YES to NO. With YES doxygen will generate short output names, but from one elastix distribution to another the names do not necessarily correspond to the same class. With NO the file names are just the same as the file names in the src dir, which is much more reproducible. Now we don't have to update links at the website all the time.