SK:
-ENH: added some cmake code to support external directories with user-defined components. I tested it and it seems to work. It uses the variable ELASTIX_USER_COMPONENT_DIRS. There you can fill in a list of directories in which cmake should search for cmakelists.txt files that include the line
  ADD_ELXCOMPONENT( componentname ....
Cmake performs a recursive search, so all subdirectories will be searched as well. Now we can make a branch in the elastix svn with private components, which we do not want to distribute. Also you can have your own local directory with experimental components. 
NB:
It is not allowed to put the component name on the next line, as in:
ADD_ELXCOMPONENT(
  componentname 
  ...)
You may however put the file names on the next lines:
ADD_ELXCOMPONENT( componentname 
  ...
  ...)