COMP: multi-threaded the CombinationImageToImageMetric
This was done using the itk::MultiThreader and for the GetValueAndDerivative() only.
It's use can be switched n and off with the (temporary, for testing) command line option -mt

The metrics were not thread-safe, however: the function SetTransformParameters and the updating of the sampler were spoiling it.
When the combo metric is used in a multi-threaded fashion, then these functions should be called once before the threading starts, and not within the metric computation anymore.
When the single-threaded version is used, for example when the normal multi-resolution registration is used, then these functions should be called within the metric computation. We cannot do it beforehand, since the optimizer just calls GetValueAndDerivative and does not do something like PreGetValueAndDerivative.
To this end we introduced a boolean m_UseMetricSingleThreaded in the AdvancedImageToImageMetric and SingleValuedPointSetToPointSetMetric classes, which controls whether these unsafe functions are called within the metric computation. Additionally, the unsafe functions are grouped in a function BeforeThreadedGetValueAndDerivative.