PERF: worked on multi-threading the MSD metric Tests show identical results with and w/o threading and a speedup
The code needs to be cleaned up. There are some booleans that allow to switch between the old single-threaded code and the new multi-threaded code.
Renamed the old GVAD to GetValueAndDerivativeSingleThreaded.

Multi-threaded samplers are turned on using "-mts true" on the command line
Multi-threaded metrics are turned on using "-mtm true" on the command line

BUG: the registrations should not set the number of threads
This causes conflicts with the number of threads defined in the metric