PERF: enhancements to the GetJacobian() of the B-spline transform.
Did some profiling with callgrind (part of valgrind) on goliath, release mode
Made enhancements to GetJacobian() of the B-spline which was measured to take 25% of the computation time (TransformPoint takes 28%) of GetValueAndDerivative().
Of that ~35% went to the fill(0) operation of the Jacobian, ~32% to ComputeNonZeroJacobianIndices(), and only 13% to the Evaluate() function that really does the work.

So:
1) I removed the fill (done for every voxel for every iteration) and replaced it with a single fill per iteration in the GetValueAndDerivative(). Currently only done for SSD.
2) I replaced the ComputeNonZeroJacobianIndices() implementation with one that avoids the use of iterators.

ENH: now measuring the time it takes to fill the derivative with zeros for initialization. memcopy is probably faster. Still to do.
ENH: Made a note to check the performance of the assignment operation in "this->m_ScaledCurrentPosition = parameters;" in itkScaledSingleValuedNonLinearOptimizer.cxx.
STYLE: some code cleanup.
ENH: avoiding the use of some member variables (m_SampleContainer and m_SampleContainerSize).