BUG/PERF: removing fill() in GetJacobian() just like that was not possible
Some parts of the Jacobian are not filled. On Windows zero was automatically substituted, resulting in a correct answer. But on Linux random values are filled in. I fixed this by checking for a B-spline transform and looping only over the nonzero entries of the Jacobian when taking the inner product with the moving image gradient. See the function EvaluateTransformJacobianInnerProduct(), which may be moved to the AdvancedImageToImageMetric. This also resulted in a measurable performance improvement.