ENH: Don't compute moving image gradient for 2D/3D registration
To avoid this computation we check if the moving image interpolator is
the RayCastInterpolator in CheckForBSplineInterpolator() in
AdvancedImageToImageMetric. This is a bit of a hack, the design used to
override the ComputeGradient function of ITK is to compute a central
difference derivative when SetComputeGradient is off. By doing this,
there is no way to ask no gradient computation at all (to save memory).
The best solution would be to override the ComputeGradient of ITK by
computing a central difference derivative. This way SetComputeGradient
will enable or disable the Gradient computation and let class that
inherit from AdvancedImageToImageMetric choses if it needs the
precomputation of the gradient.

For more details see the post about "2D/3D registration memory issue" in
elastix's mailing list (2 July 2012)