PERF: implemented derivatives for the linear interpolator
This can replace the first order B-spline interpolator. Previously, the linear interpolator did not have the derivative and only the value. This forced us to use the first order B-spline. That one however has a high computational complexity. The linear interpolator was about 10x-15x faster on my machine, when used standalone. In a registration setting it saved about 10-15%.

Note that the advanced linear interpolator implements mirroring boundary conditions, just like the B-spline, but different from the original linear interpolator (which uses a constant boundary condition).

ENH: A test for the advanced linear interpolator was added.

PERF: Compute the value and gradient of the moving image in one go, instead of two, saving moving image accesses (GetPixel()). Also benefits the B-spline interpolator.