PERF: Created an optimized version of GetJacobianOfSpatialHessian()
A lot of time was (and still is) needed for taking into account the direction cosines and grid spacing.
Code was sped up by doing that less, but it is still an expensive operation.

Code was further sped up for the case that the m_PointToIndexMatrix is a diagonal matrix.

The itkAdvancedBSplineDeformableTransformTest.cxx revealed a 5-fold speed-up.