PERF: using the scanline iterator at several places
The ImageScanlineIterator is a recent addition to the ITK by Bradley Lowekamp, with a much cheaper operator++. It does not do bounds checking. When applied to the B-spline TransformPoint() the registration time reduced by almost 10% in some cases.

Currently, I have implemented it for the B-spline: TransformPoint, GetSpatialJacobian, GetSpatialHessian, GetJacobianOfSpatialJacobian, GetJacobianOfSpatialHessian; and for the MI: ComputeValueAndPRatioArray, UpdateJointPDFAndDerivatives, NormalizeJointPDF, NormalizeJointPDFDerivatives, AfterThreadedComputePDFs.


STYLE: removed some unused code