This is a metric based on the T1 mapping model, using 3 model parameters, C1, C2 and C3: I(x) = C1 * exp( -t * C2 ) + C3 t is the time that the image is recorded, the so called trigger time. The triggertimes must be added to the parameterfile.
An optimizer is used to find C1, C2 and C3, which can be extended to a Levenberg-
Marquardt optimizer by setting some parameters. 

The metric is defined as sum_x(I(x) - M(x))^2, where x is a possible voxel location
in both space and time. I is the estimated model image (see definition above) 
and its derivative w.r.t. mu is taken to be constant.

The ModelBaseClass is a class in which the virtual functions for a model is defined.
The T1MappingModel defines the T1 mapping model. It has an evaluate() and an
evaluateDerivative() function. In the evaluateDerivative() function, the derivatives
of the model w.r.t. the model parameters is defined. These are necessary for the
optimizer.
The LMOptimizer is a Levenberg-Marquardt optimizer. Values for the 
number of iterations, the damping parameter and the scaling factor can be 
set there.