COMP: ITK started using a FixedParameters type
Before this patch elastix compiled with ITK 4.7.x and ITK 4.8, but not with ITK git later than ITK 4.8.
After this patch elastix compiles with ITK 4.8 and later, but not with ITK 4.7.x anymore.

The compilation problem was reported by Linqing Feng from the Korean Institute of Science and Technology, who also provided an initial patch.