ENH: updating test results after latest changes
There are some minor changes in the results after commit #5349.

The two likely candidates are the change in BSplineDerivativeKernelFunction2, where a bug was fixed for the first order B-spline interpolator, and BSplineKernelFunction2, where a small numerical change was made in the third order B-spline.

A few, but not all tests, slightly changed their results.