ENH: removing testing dependency on ITKTools
We previously depended on pxcastconvert and pxcomputeoverlap from
ITKTools to be installed on the path. This is troublesome for systems
where these programs were not installed, for example the CI systems we
use for continuous testing.

The dependency on pxcastconvert was completely removed and replaced by
the python command os.replace().

A stripped down version of pxcomputeoverlap is now integrated in the
repository, and build with elastix.

The path to the build directory is now given as a command line option to
the evaluation scripts (python), using CMake generator expressions
(thank you Kasper).

All tests were verified to run on my local Windows machine.