BUG: No longer read rotation center as index (elastix transform < 3.402)
No longer read "CenterOfRotation" from an old transform parameter file, which is the rotation center of the transformation, specified by its index coordinates.

elastix version >= 3.402 no longer specifies the rotation center by its index coordinates.

As explained by Stefan @stefanklein

"Having a center of rotation in index coordinates in the TransformParameter file is bad practice, since its conversion to a world coordinate depends on the knowledge of the size/spacing/origin/directioncosines of the fixed image that was used for registration. That information is stored in the TransformParameter file as well, but it is easy to forget changing the CenterOfRotationIndex when adjusting the (fixed image) resample domain."

Also discussed with Marius @mstaring

Fixes issue #162 ("AffineLogStackTransform: uninitialized local variable 'rdorigin' used")