ENH: Add CMake, _MSC_BUILD, and ITK_VERSION_PATCH to --extended-version
Extended the output of the elastix command-line option "--extended-version" by adding:

 * The CMake version number
 * _MSC_BUILD: the revision number element of MSVC's version number
 * The patch component of ITK's version number

It appeared that `ITK_VERSION_STRING` only shows the major and the minor version component, for example, "5.0". With this commit, the patch number is included, as in "5.0.1".

The CMake version number is explicitly mentioned at https://github.com/SuperElastix/elastix/releases so it seems useful to provide it through command-line option "--extended-version" as well.