DOC: Fix compilation of code examples "Using elastix as a library"
ITK5 no longer supports the `int` value 0 as initial value of `SmartPointer`, so it is replaced by `nullptr` in the code examples.

`TRANSFORMIX::TransformImage` has an extra output_path parameter that was missing.

Placed `ELASTIX` and `TRANSFORMIX` on the stack, instead of the heap, getting rid of unnecessary code to explicitly clean up memory.