COMP: Fix build error on MacOS by using vnl_matrix_fixed
Fixes the error:
/.../SlicerElastix-build/elastix/Common/Transforms/itkAdvancedRigid2DTransform.hxx:125:7: error: no viable conversion from 'const InternalMatrixType' (aka 'const vnl_matrix_fixed<double, 2U, 2U>') to 'vnl_matrix<double>'

Note that `itk::Matrix<T,NRows,NColumns>::GetVnlMatrix()` returns a reference to a `vnl_matrix_fixed<T,NRows,NColumns>` (not to a `vnl_matrix<T>`), and is doing so for many years already. So it appears preferable to store the result of the function call in a "fixed" matrix type anyway.

Also replaced `vnl_svd<TScalarType>` by `vnl_svd_fixed<TScalarType,2,2>`.