ADD: Addition of point to surface metric to Elastix.
COMP: Remove cmake* from .gitignore

Having cmake* in there stopped Gokhan Gunay from adding his CMakeLists file to pull request #236 (Addition of PointToSurface metric).

Addition of PointToSurface metric

ADD: Addition of point to surface metric to Elastix.

In this commit addition of point to surface metric is done.
This metric essentially considers how surface of an masked organ
is aligned with the points provided from the surface of the organ.
Therefore, it is used to improve alignment of the organ under interest.