STYLE: Remove ViolaWellsMutualInformation from Metrics components
This component appears obsolete, and may be removed from the repository, as was agreed upon during our internal Elastix sprint (with Stefan Klein and Marius Staring).

Note that this component was OFF by default already.

Also removed obsolete comment about metrics that "do not support the image sampler functionality", which appeared to apply specifically to ViolaWellsMutualInformation.