STYLE: Simply define TransformBase::ParametersType = OptimizerParameters
It appears unnecessary to make the definition of `TransformBase<TElastix>::ParametersType` indirectly dependent on `TElastix`. This commit simply defines `ParametersType` as `itk::OptimizerParameters<ValueType>`, with `ValueType = double`. Which was what always turned out to be when everything was evaluated.

This commit aims to improve code readability, and eventually also compilation speed.