ENH: Elastix library should read the initial transform from file
When the name of the initial transform is not "NoInitialTransform", and it cannot be interpreted as an index value (unsigned integer), the Elastix library should just assume that it is the file name of the transform file, and read it, in `TransformBase::ReadFromFile()`.

The elastix library and the elastix executable should deal with this in the same way.

As discussed with Viktor van der Valk and Marius Staring.