COMP: Replace double underscores from include guards by "itk"
Follows the ITK coding convention: The names of include guards of "itk*.h" and "itk*.hxx" files always start with "itk". Moreover, it is also according to the C++ Standard, which says that each identifier that contains a double underscore `__` is reserved to the compiler and the Standard Library implementation.