STYLE: Remove space between class and member names in C++ source files
It appears that clang-format has in some cases introduced a space character between `ClassName` and `::MemberName`. This typically seems to have happened when there originally was a line break between the class name and the member name.

This commit fixes the formatting by removing those spaces, as follows:

    find . \( -iname *.cxx \) -exec perl -pi -w -e 's/([A-Z][a-z]+) \:\:([A-Z][a-z])/$1::$2/g;' {} \;

Follows ITK:
 - pull request https://github.com/InsightSoftwareConsortium/ITK/pull/2096
 - commit: https://github.com/InsightSoftwareConsortium/ITK/commit/734d6f84d728b343baae5ef66643a5568a5a30c7