COMP: Remove itk::InvalidImageMomentsError from Calculator hxx file
The class `itk::InvalidImageMomentsError` appears unused, while it has been defined twice:

 - https://github.com/InsightSoftwareConsortium/ITK/blob/v5.1.1/Modules/Filtering/ImageStatistics/include/itkImageMomentsCalculator.hxx#L28
 - https://github.com/SuperElastix/elastix/blob/5.0.1/Common/Transforms/itkAdvancedImageMomentsCalculator.hxx#L29

Which caused compilation errors like:

> ITK511\Modules\Filtering\ImageStatistics\include\itkImageMomentsCalculator.hxx(28,32): error C2011: 'itk::InvalidImageMomentsError': 'class' type redefinition
> elx\elastix\Common\Transforms\itkAdvancedImageMomentsCalculator.hxx(29): message : see declaration of 'itk::InvalidImageMomentsError'

When doing:

    #include "AdvancedAffineTransform/elxAdvancedAffineTransform.h"
    #include "EulerTransform/elxEulerTransform.h"

Note that hxx files are meant for implementation only. A class defined within an hxx file should not be not be part of the interface (API), otherwise it should be defined in the h file instead.