ENH: DeformationFieldFileName when "TransformParameters." is not found
When the current transform parameter filename would not have "TransformParameters." as a substring, the creation of the "lastpart" of a `DeformationFieldFileName` would go wrong (typically causing an `std::out_of_range` exception). However, it still appears useful to make a `DeformationFieldFileName` in that case. Especially when trying to write a default-constructed transform, for test purposes.

This commit makes the "lastpart" empty, in that case. Discussed with Marius at the weekly internal elastix sprint.