STYLE: Avoid conversions between (signed/unsigned) char and string
So far, it appears unnecessary to support 8-bits (signed or unsigned) `char` as type of a parameter values. Moreover, conversion between `char` and `std::string` appears rather tricky, as it may not always be obvious whether to use the character value or the integer value represented by the bits of a `char`.

Simplified `ParameterMapInterface::StringCast` by removing the `char` specific `NumericTraits<T>::AccumulateType` workaround.