ENH: Replace itkGPUKernelClassMacro calls by GetOpenCLSourceOf functions
Removed all (n = 22) `itkGPUKernelClassMacro` macro calls, as there is an upcoming breaking change of the macro definition with ITK 5.2 (introduced with ITK v5.2rc01).

Replaced the class definitions that were created by the macro by simple functions of the form `GetOpenCLSourceOf...Kernel()`, which just return the OpenCL source code as a string.

Allows compilation on ITK 5.1.1 and 5.1.2, as well as ITK 5.2.0 (release candidate 1 or 2).