BUG: ReadParameter should fail if parameter string isn't fully processed
The `ReadParameter<T>` member functions of `ParameterMapInterface` should fail when a parameter string is not entirely processed, during the conversion from `std::string` to type `T`.

For example, when the parameter has string value "0.1", an attempt to read it into an integer will only process the string until (just before) the decimal point. This should trigger an exception.

Note that a string value like "123.000" should still be converted to an integer, by `ReadParameter<int>`, as it only has trailing zero's after the decimal point.