STYLE: Replace `std::string::compare` calls by simple `==` expressions
Ran Clang-Tidy-Fix (LLVM 11.1.0) `readability-string-compare`:

https://clang.llvm.org/extra/clang-tidy/checks/readability-string-compare.html

Which says:

"A common mistake is to use the string’s compare method instead of using the equality or inequality operators. The compare method is intended for sorting functions and thus returns a negative number, a positive number or zero depending on the lexicographical relationship between the strings compared. If an equality or inequality check can suffice, that is recommended. This is recommended to avoid the risk of incorrect interpretation of the return value and to simplify the code."

Fixes Clang warnings like:

> elxResamplerBase.hxx(501): warning: do not use 'compare' to test equality of strings; use the string equality operator instead [readability-string-compare]