ENH: Add support for multiple metrics and regularization
In addition to m_NumberOfMetrics, three counters were added to the itkCombinationImageToImageMetric: m_NumberOfImageMetrics, m_NumberOfTransformMetrics and m_NumberOfPointSetMetrics.
Casting is used to verify the type of metric and to count metrics independetly.
The number of interpolators, images, masks and image pyramids is checked only against m_NumberOfMetrics.
The transform and point set metric counters may be useful in the future.

resolves #481