STYLE: Concatenate "literal" strings, consecutively inserted into stream
Concatenated literal strings that were consecutively inserted into an output stream, to reduce overhead in compilation and runtime performance, and to simplify the code.

An exception is made for a pair of consecutively inserted literal string for which the first one ends with `\n`.

By Notepad++ v8.1.4, Find in Files:

    Find what: ([^\\][^n])"\r\n[ ]+<< "
    Replace with: $1
    Filter: *.hxx;*.h;*.cxx
    [v] Match case
    (*) Regular expression