STYLE: Simply declare _classname##_install template parameter `unsigned`
The template is specialized for `NrOfSupportedImageTypes + 1`, which is also just an unsigned int. Using the typedef `::elastix::ComponentDatabase::IndexType` in this case appears somewhat overdone. Note that NrOfSupportedImageTypes is `5` by default, and it is rarely much larger. So there is no need to use anything larger than a simple unsigned int.