STYLE: Make StackTransform Parameters empty by default
Instead of using `OutputSpaceDimension` as initial "guess" for the number of parameters to be allocated, just make it zero, in the `StackTransform` default-constructor.

Eventually, the number of parameters of a stack transform is defined as the number of subtransforms _times_ the number of parameters of a single subtransform.

The original initial number (`OutputSpaceDimension`) just caused unnecessary memory reallocations.