BUG: `ConvertToSingleItkTransform` should not convert a stack transform
It appears that `TransformIO::ConvertItkTransformBaseToSingleItkTransform` would accidentally convert a stack transform to its own type (making an unnecessary copy), if the stack transform type was registered to the `itk::TransformFactory`.

This mistake is avoided here, by ensuring that during conversion, the class name of the target transform differs from the original transform.