STYLE: Remove m_MaxNumberInterpolationPoints from Reduced ImageFunction
Removed the private data member `m_MaxNumberInterpolationPoints` from `ReducedDimensionBSplineInterpolateImageFunction`. It appears unnecessary, because it was always equal to the number of elements of `m_PointsToIndex` anyway.

Used a range-based for-loop to iterate over `m_PointsToIndex` in both `EvaluateAtContinuousIndex`, and `EvaluateDerivativeAtContinuousIndex`, which might be slightly faster that the original old-style for-loop.