PERF: Avoid redundant CPU call when BSplineInterpolator
The interpolator is always being called before the GenerateData() of the OpenCLResampler. For only the case of the BSplineInterpolator, this meant
that it was doing real work in the CPU via the BSplineDecompositionImageFilter. This work was then repeated when the GPUResampler that belongs to
OpenCLResampler called its own GenerateData(). Now the default LinearInterpolator is NOT overwritten when OpenCLResampler is properly setup.