BUG: Avoid crash when m_Elastix->GetFixedImage() returns nullptr
Instead of using the pointer returned by `GetFixedImage()` directly, pass it to `elx::Deref(ptr)`, which throws an exception if it is null. Avoids a common case of undefined behavior, when the actual pixel type of the user specified fixed image was not included with the appropriate  `ELASTIX_IMAGE_nD_PIXELTYPES` CMake parameter. Allows finding out about the mistake in a more gentle way.