PERF: Make EvaluateParzenValues calls faster, using raw buffer of values
Added a `ParzenWindowHistogramImageToImageMetric::EvaluateParzenValues` overload that has a raw pointer to the buffer of values as parameter.

Stored the local Parzen values of both fixed and moving image together in one array, instead of having separate arrays for fixed and moving images, in both ParzenWindowMutualInformationImageToImageMetric::UpdateDerivativeLowMemory and ParzenWindowHistogramImageToImageMetric::UpdateJointPDFAndDerivatives.

A performance improvement of ~5% was observed, executing an ElastixRegistrationMethod on VS2019 Release, using the default registration parameter maps.