BUG: Library should write to output directory whenever it is specified
Whenever the output directory is specified, the library will now write files to this directory, just like the executable does.

(Before this commit, the elastix library would only create the result images in memory, even when the WriteResultImage parameter would be "true". This was an inconsistency between the library and the elastix executable.)

With this commit, an empty string (`""`) is used internally to indicate that the output directory is unspecified. The use of a magical "output_path_not_set" string is abandoned.