STYLE: Let GetNthTransform, GetCombinationTransform return a raw pointer
In general, the use of a smart pointer as return type allows extending the lifetime of the returned object beyond the function call. But as these two `ElastixRegistrationMethod` member functions both just return a pointer to a transform object that is "owned" by the `ElastixRegistrationMethod` object, the caller of `GetNthTransform` or `GetCombinationTransform` does not need to extend the lifetime of the returned transform object. This returned transform object will stay alive while the `ElastixRegistrationMethod` object is there, even when the return type is a raw pointer.

This adjustment was already proposed by Matt McCormick.