PERF: Optimize FullSampler for if input image and mask have same domain
Similar to the existing optimization in itkComputeImageExtremaFilter.hxx

When the mask and the image have exactly the same image domain, the sampler may become more than 3x as fast as before, from more than 0.60 sec. (before this commit) to less than 0.18 (after this commit), on an image of 4096x4096 pixels, using VS2022 Release.