STYLE: Replace `DerivativeValueType{}` with `0.0`
`DerivativeValueType` is just a floating point type. In both AdvancedImageToImageMetric and SingleValuedPointSetToPointSetMetric, it is defined as:

    using DerivativeValueType = typename DerivativeType::ValueType;

In ITK 5.4's SingleValuedCostFunction, DerivativeType is defined as:

    using DerivativeType = Array<ParametersValueType>;

SingleValuedCostFunction is derived from CostFunction, which has:

    using ParametersValueType = TInternalComputationValueType;

And for CostFunction, TInternalComputationValueType = double.

In LimiterFunctionBase, DerivativeValueType is defined as RealType:

      using DerivativeValueType = OutputType;

OutputType is an alias of NumericTraits<TInput>::RealType