STYLE: Avoid copying `ImpactModelConfiguration`, let `GetModel()` return a reference (#5)
* STYLE: Avoid copying `ImpactModelConfiguration` objects

Multiple copies of an `ImpactModelConfiguration` would share the same model, which could be confusing, especially when the model is being modified.

It might also be slightly more efficient to pass a `ImpactModelConfiguration` object "by const reference", instead of "by value".

* STYLE: Let GetModel() return a reference, instead of a shared_ptr

Aims to simplify the interface of `ImpactModelConfiguration`.

Aims to express that the model returned by GetModel() is owned and managed by an `ImpactModelConfiguration`. So the lifetime of the model should not be extended beyond the lifetime of the `ImpactModelConfiguration`.