ENH: Fixed width types ("int8", "float64", etc) for ResultImagePixelType
Supported any of the following values for the elastix/transformix parameter `ResultImagePixelType`:

  "int8", "uint8", "int16", "uint16", "int32", "uint32", "int64", "uint64", "float32", "float64".

The number of bits of such a fixed width type is guaranteed to be the same on all supported platforms.

Note that this commit adds support for 64-bit integer types on Windows (which was not supported on Windows before, because its "long" is only 32 bits).